added common file reading elements in datasets_utils
