+++ torchtext/datasets/yelpreviewpolarity.py
@@ -1,11 +1,13 @@
-from torchtext.utils import download_from_url, extract_archive, unicode_csv_reader
-from torchtext.data.datasets_utils import _RawTextIterableDataset
-from torchtext.data.datasets_utils import _wrap_split_argument
-from torchtext.data.datasets_utils import _add_docstring_header
-from torchtext.data.datasets_utils import _find_match
 import os
-import io
-from torchtext.data.datasets_utils import _create_dataset_directory
+from torchtext.utils import download_from_url, extract_archive
+from torchtext.data.datasets_utils import (
+    _RawTextIterableDataset,
+    _wrap_split_argument,
+    _add_docstring_header,
+    _find_match,
+    _create_dataset_directory,
+    _create_data_from_csv,
+)
 
 URL = 'https://drive.google.com/uc?export=download&id=0Bz8a_Dbh9QhbNUpYQ2N3SGlFaDg'
 
@@ -23,11 +25,6 @@ DATASET_NAME = "YelpReviewPolarity"
 @_create_dataset_directory(dataset_name=DATASET_NAME)
 @_wrap_split_argument(('train', 'test'))
 def YelpReviewPolarity(root, split):
-    def _create_data_from_csv(data_path):
-        with io.open(data_path, encoding="utf8") as f:
-            reader = unicode_csv_reader(f)
-            for row in reader:
-                yield int(row[0]), ' '.join(row[1:])
     dataset_tar = download_from_url(URL, root=root,
                                     path=os.path.join(root, _PATH),
                                     hash_value=MD5, hash_type='md5')
